+++ config/universal-webpack-settings.js
@@ -1,0 +1,8 @@
+export default
+{
+  server:
+  {
+    input: './server/server.js',
+    output: './build/server/server.js'
+  }
+}